backend/requirements.txt
fastapi
uvicorn[standard]
langgraph
langgraph-checkpoint-sqlite
anthropic
pydantic
pydantic-settings
python-dotenv
httpx
psycopg2-binary
thefuzz